sntp: use one time source and fix unsynchronized sntp time stamp
