+++ cores/esp8266/coredecls.h
@@ -12,9 +12,6 @@ extern "C" {
 #include <stdint.h>
 #include <cont.h> // g_pcont declaration
 
-extern bool timeshift64_is_set;
-extern uint32_t sntp_real_timestamp;
-
 bool can_yield();
 void esp_yield();
 void esp_schedule();
