--- cores/esp8266/sntp-lwip2.h
@@ -1,6 +1,0 @@
-#ifndef __sntp_lwip2_h__
-#define __sntp_lwip2_h__
-
-#include <coredecls.h>
-
-#endif
